Moving exception to common
