+++ api/src/main/kotlin/com/example/todo/domain/service/TaskService.kt
@@ -2,7 +2,7 @@ package com.example.todo.domain.service
 
 import com.example.todo.domain.entity.Task
 import com.example.todo.domain.entity.TaskEntity
-import com.example.todo.domain.exception.NotFoundException
+import com.example.todo.exceptions.NotFoundException
 import com.example.todo.infrastructure.mapper.TaskMapper
 import org.springframework.stereotype.Service
